+++ QAction_1/API/Repositories/GetRepositoryReleases.cs
@@ -12,7 +12,7 @@
         public string Url { get; set; }
 
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("node_id")]
         public string NodeId { get; set; }
@@ -54,7 +54,7 @@
         public string Login { get; set; }
 
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("node_id")]
         public string NodeId { get; set; }
@@ -168,7 +168,7 @@
         public string Login { get; set; }
 
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("node_id")]
         public string NodeId { get; set; }
